Restructure fixtures, pages, journeys and tests merged from master
