--- e2e/fixtures/citizen/caseView/respondent/contactPreferences/confirmationContent.ts
@@ -1,4 +1,0 @@
-export enum ConfirmationContent {
-  GovukHeadingL = "Contact preferences updated",
-  GovukBodyM = "You will receive updates on the case by post.",
-}
