--- e2e/fixtures/citizen/caseView/respondent/contactPreferences/contactPreferencesContent.ts
@@ -1,15 +1,0 @@
-export enum ContactPreferencesContent {
-  GovukHeadingXL = "Contact Preferences",
-  GovukBody1 = "You can choose to receive case updates by email or post.",
-  GovukBody2 = "If you receive updates by email, the updates will also be available to view in your dashboard.",
-  GovukBody3 = "This includes updates on:",
-  li1 = "court orders",
-  li2 = "hearings",
-  li3 = "decisions in your case",
-  GovukFieldsetLegend = "How would you prefer to be contacted?",
-  govukHint1 = "Select one of these options.",
-  govukHint2 = "All communication from the court will be sent by email.",
-  govukHint3 = "All communication from the court will be sent by post.",
-  digital = "Digital",
-  byPost = "Post",
-}
