--- e2e/fixtures/citizen/caseView/respondent/contactPreferences/reviewContent.ts
@@ -1,11 +1,0 @@
-export enum ReviewContent {
-  GovukHeadingXL = "Contact Preferences",
-  GovukHeadingL = "Personal details",
-  GovukSummaryListKey = "Address",
-  GovukBody1 = "Heaven",
-  GovukBody2 = "7 Castle Square",
-  GovukBody3 = "Swansea",
-  GovukBody4 = "SA1 1DW",
-  GovukBody5 = "You have decided to receive updates by post.",
-  GovukBody6 = "You will no longer receive updates by email. You can still access previous updates through your dashboard.",
-}
