+++ e2e/pages/citizen/caseView/contactPreferences/confirmationPage.ts
@@ -1,16 +1,14 @@
 import { Page } from "@playwright/test";
-import { Selectors } from "../../../../../common/selectors.ts";
-import { CommonStaticText } from "../../../../../common/commonStaticText.ts";
-import AccessibilityTestHelper from "../../../../../common/accessibilityTestHelper.ts";
-import {
-  ConfirmationContent
-} from "../../../../../fixtures/citizen/caseView/respondent/contactPreferences/confirmationContent.ts";
-import { Helpers } from "../../../../../common/helpers.ts";
+import { Selectors } from "../../../../common/selectors.ts";
+import { CommonStaticText } from "../../../../common/commonStaticText.ts";
+import AccessibilityTestHelper from "../../../../common/accessibilityTestHelper.ts";
+import { ConfirmationContent } from "../../../../fixtures/citizen/caseView/contactPreferences/confirmationContent.ts";
+import { Helpers } from "../../../../common/helpers.ts";
 
 export class ConfirmationPage {
   public static async confirmationPage(
     page: Page,
-    accessibilityTest: boolean
+    accessibilityTest: boolean,
   ): Promise<void> {
     // this part of the case is complete when the case is created through courtnav so only need to check the page
     await this.checkPageLoads(page, accessibilityTest);
@@ -19,19 +17,19 @@ export class ConfirmationPage {
 
   private static async checkPageLoads(
     page: Page,
-    accessibilityTest: boolean
+    accessibilityTest: boolean,
   ): Promise<void> {
     await page
       .locator(Selectors.GovukHeadingL, {
-        hasText: ConfirmationContent.GovukHeadingL
+        hasText: ConfirmationContent.GovukHeadingL,
       })
       .waitFor();
     await Promise.all([
       Helpers.checkVisibleAndPresent(
         page,
         `${Selectors.GovukBodyM}:text-is("${ConfirmationContent.GovukBodyM}")`,
-        1
-      )
+        1,
+      ),
     ]);
     if (accessibilityTest) {
       await AccessibilityTestHelper.run(page);
@@ -40,7 +38,7 @@ export class ConfirmationPage {
 
   private static async saveAndContinue(page: Page): Promise<void> {
     await page.click(
-      `${Selectors.GovukButton}:text-is("${CommonStaticText.continue}")`
+      `${Selectors.GovukButton}:text-is("${CommonStaticText.continue}")`,
     );
   }
 }
