--- e2e/pages/citizen/caseView/respondent/contactPreferences/reviewPage.ts
@@ -1,56 +1,0 @@
-import { Page } from "@playwright/test";
-import { Selectors } from "../../../../../common/selectors.ts";
-import { CommonStaticText } from "../../../../../common/commonStaticText.ts";
-import AccessibilityTestHelper from "../../../../../common/accessibilityTestHelper.ts";
-import { ReviewContent } from "../../../../../fixtures/citizen/caseView/respondent/contactPreferences/reviewContent.ts";
-import { Helpers } from "../../../../../common/helpers.ts";
-
-export class ReviewPage {
-  public static async reviewPage(
-    page: Page,
-    accessibilityTest: boolean
-  ): Promise<void> {
-    // this part of the case is complete when the case is created through courtnav so only need to check the page
-    await this.checkPageLoads(page, accessibilityTest);
-    await this.saveAndContinue(page);
-  }
-
-  private static async checkPageLoads(
-    page: Page,
-    accessibilityTest: boolean
-  ): Promise<void> {
-    await page
-      .locator(Selectors.GovukHeadingXL, {
-        hasText: ReviewContent.GovukHeadingXL
-      })
-      .waitFor();
-    await Promise.all([
-      Helpers.checkVisibleAndPresent(
-        page,
-        `${Selectors.GovukHeadingL}:text-is("${ReviewContent.GovukHeadingL}")`,
-        1
-      ),
-      Helpers.checkVisibleAndPresent(
-        page,
-        `${Selectors.GovukSummaryListKey}:text-is("${ReviewContent.GovukSummaryListKey}")`,
-        1
-      ),
-      Helpers.checkGroup(
-        page,
-        6,
-        ReviewContent,
-        `GovukBody`,
-        `${Selectors.GovukBody}`
-      )
-    ]);
-    if (accessibilityTest) {
-      await AccessibilityTestHelper.run(page);
-    }
-  }
-
-  private static async saveAndContinue(page: Page): Promise<void> {
-    await page.click(
-      `${Selectors.GovukButton}:text-is("${CommonStaticText.continue}")`
-    );
-  }
-}
